fix: git 에러 해결 및 기능복구 70% 완료

+++ src/pages/join/join_page.tsx
@@ -2,26 +2,39 @@ import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import Wrapper from '../../components/common/wrapper';
 import { stringLiteral } from '@babel/types';
-import { Routes, Route, Link } from 'react-router-dom';
+import { Routes, Route, Link, useNavigate } from 'react-router-dom';
+
 
 //더미데이터 테스트용
-const User = {
-  email: 'test@test.com',
-  password: 'qwer1234!!',
-};
+const User = [
+  {
+    email: 'test@test.com',
+    password: 'qwer1234!!',
+  },
+  {
+    email: '1004@test.com',
+    password: '1004qwer!!',
+  },
+];
+
+
 
 const JoinPage = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [confirmPw, setConfirmPw] = useState('');
-  const [name, setName] = useState('');
-  const [number, setNumber] = useState('');
+  const [email, setEmail] = useState(''); // 이메일 상태, 상태변경
+  const [password, setPassword] = useState(''); // 비밀번호 상태, 상태변경
+  const [confirmPw, setConfirmPw] = useState(''); // 비밀번호 재입력 상태, 상태변경
+  const [name, setName] = useState(''); // 이름 상태, 상태변경
+  const [number, setNumber] = useState(''); // 전화번호 상태, 상태변경
+  
 
-  const [emailValid, setEmailValid] = useState(false);
-  const [passwordValid, setPasswordValid] = useState(false);
-  const [ConfirmPwValid, setConfirmPwValid] = useState(false);
-  const [ConfirmNameValid, setConfirmNameValid] = useState(false);
+  const [emailValid, setEmailValid] = useState(false); // 이메일 벨리데이션 유효성 검사
+  const [passwordValid, setPasswordValid] = useState(false); // 비밀번호 벨리데이션 유효성 검사
+  const [ConfirmPwValid, setConfirmPwValid] = useState(false); // 비밀번호 재입력 벨리데이션 유효성 검사
+  const [ConfirmNameValid, setConfirmNameValid] = useState(false); // 이름 벨리데이션 유효성 검사
+  const [numberValid, setNumberValid] = useState(false); // 전화번호 유효성 검사
   const [notAllow, setNotAllow] = useState(true); //이메일, 패스워드가 맞게 작동할때 버튼이 활성화 되는 기능
+
+  const navigate  = useNavigate();// 회원가입 페이지에서 회원가입 버튼 누르고 성공할 시 로그인 페이지로 넘어가는 기능
 
   //submit 버튼 활성화 기능 구현 코드
   useEffect(() => {
@@ -81,12 +94,11 @@ const JoinPage = () => {
   //이름 validation 유효성 검사
   const handleName = (e: React.ChangeEvent<HTMLInputElement>) => {
     setName(e.target.value);
-
     {
       /* 이름 valid 유효성 검사 */
     }
-
-    if (name.length === 3) {
+    const regex = /^[가-힣]{3,4}$/
+    if (regex.test(name)) {
       setConfirmNameValid(true);
     } else {
       setConfirmNameValid(false);
@@ -95,22 +107,25 @@ const JoinPage = () => {
 
   //이름 validation 유효성 검사
   const handleNumber = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setNumber(e.target.value);
+    const numberValue = e.target.value;
+    setNumber(numberValue);
+    {
+      /* 이름 valid 유효성 검사 */
+    }
 
-    // {
-    //   /* 이름 valid 유효성 검사 */
-    // }
+    const regex = /^\d{11}$/;
 
-    // if (name.length === 3) {
-    //   setConfirmNameValid(true);
-    // } else {
-    //   setConfirmNameValid(false);
-    // }
+    if (regex.test(numberValue)) {
+      setNumberValid(true);
+    } else {
+      setNumberValid(false);
+    }
   };
 
   //이메일 입력 후 중복확인 버튼 눌렀을때 맞게 되었는지 확인창 기능
   const onClickConfirmBtn = () => {
-    if (email === User.email) {
+    const foundUser = User.find(userData => userData.email === email)
+    if (foundUser) {
       alert('사용가능한 이메일입니다.');
     } else {
       alert('사용할 수 없는 이메일입니다.');
@@ -119,10 +134,23 @@ const JoinPage = () => {
 
   //가입하기 버튼 로컬스토리지.setItem
   const handleJoin = () => {
-    localStorage.setItem('email', email);
-    localStorage.setItem('password', password);
+    if (
+      !emailValid ||
+      !passwordValid ||
+      !ConfirmPwValid ||
+      !ConfirmNameValid ||
+      !numberValid
+    ) {
+      alert("모든 항목을 입력해주세요.");
+    } else {
+      // 모든 유효성 검사 통과한 경우
+      localStorage.setItem('email', email);
+      localStorage.setItem('password', password);
+      
+      alert("회원가입에 성공하였습니다.");
 
-    alert("회원가입에 성공하였습니다.")
+      navigate('/loginpage'); // 가입 성공 시 로그인 페이지로 이동
+    }
   }
 
   
@@ -166,7 +194,7 @@ const JoinPage = () => {
           </ErrorMessageWrap>
 
           {/* 비밀번호 입력칸 */}
-          <InputTitle style={{ marginTop: '20px' }}>비밀번호</InputTitle>
+          <InputTitle style={{ marginTop: '26px' }}>비밀번호</InputTitle>
           <InputWrap>
             <UserIcon>
               <a
@@ -229,31 +257,37 @@ const JoinPage = () => {
               placeholder="이름을 입력해주세요."
             />
           </InputWrap>
-          {/* 비밀번호 에러메세지 입력칸 */}
+          {/* 이름 에러메세지 입력칸 */}
           <ErrorMessageWrap>
             {!ConfirmNameValid && name.length > 0 && (
               <div>비밀번호가 일치하지 않습니다.</div>
             )}
           </ErrorMessageWrap>
 
-          {/* 이름 입력칸 */}
+          {/* 전화번호 입력칸 */}
           <InputTitle style={{ marginTop: '26px' }}>전화번호</InputTitle>
           <InputWrap>
             <Input2
-              type="tel"
+              type="text"
               value={number}
               onChange={handleNumber}
               placeholder="전화번호를 입력해주세요."
             />
           </InputWrap>
+          <ErrorMessageWrap>
+            {!numberValid && number.length > 0 && (
+              <div>전화번호가 올바르지 않습니다.</div>
+            )}
+          </ErrorMessageWrap>
+
         </EmailPW>
 
         {/* 가입하기, 가입취소 버튼 */}
         <BtnWrap>
+          <LoginBtn onClick={() => handleJoin()}>가입하기</LoginBtn>
           <Link to="/loginpage">
             <Cancel>가입취소</Cancel>
           </Link>
-          <LoginBtn onClick={() => handleJoin()}>가입하기</LoginBtn>
         </BtnWrap>
       </Body>
     </Wrapper>
@@ -272,7 +306,6 @@ const Body = styled.div`
 const JoinText = styled.p`
   font-size: 30px;
   font-weight: 700;
-  text-align:center;
 `;
 
 //회원가입 부가설명
@@ -280,7 +313,6 @@ const JoinText1 = styled.p`
   font-size: 14px;
   margin-top: 10px;
   margin-bottom: 40px;
-  text-align: center;
 `;
 
 // 이메일 주소, 비밀번호 전체 틀
@@ -292,10 +324,9 @@ const EmailPW = styled.div`
 
 // 이메일주소 , 비밀번호
 const InputTitle = styled.div`
-  font-size: 16px;
-  font-weight: 500;
+  font-size: 18px;
+  font-weight: 600;
   color: #262626;
-  padding-bottom:6px;
 `;
 
 //이메일, 비밀번호 입력창 틀
@@ -308,7 +339,7 @@ const InputWrap = styled.div`
   padding: 8px;
   border: 1px solid #e2e0e0;
   background-color: white;
-  border-radius:10px;
+  box-shadow: 1px 1px 0.5px 0.5px gray;
 
   &:focus-within {
     border: 1.5px solid #339af0;
@@ -357,7 +388,7 @@ const ConfrimBtn = styled.button`
   cursor: pointer;
 
   &:disabled {
-    background-color: #a1a1a1;
+    background-color: #dadada;
     color: white;
   }
 `;
@@ -385,30 +416,33 @@ const ErrorMessageWrap = styled.div`
 const BtnWrap = styled.div`
   display: flex;
   justify-content: center;
-  margin-top:45px;
 `;
 
 const LoginBtn = styled.button`
-  border-radius: 10px;
-  padding: 18px 52px;
-  border: none;
-  cursor:pointer;
-  font-size:18px;
+  border-radius: 3px;
+  width: 200px;
+  padding: 10px;
+  border: 1px solid #e2e0e0;
+  border-radius: 6px;
   background-color: #339af0;
+  font-size: 20px;
   color: white;
-  margin-left:30px;
-  font-weight:600;
+  font-weight: 400;
+  cursor: pointer;
+  margin-top: 30px;
 `;
 
 const Cancel = styled.button`
-  background-color: #e0e0e0;
-  border-radius: 10px;
-  padding: 18px 52px;
-  border: none;
-  cursor:pointer;
-  font-size:18px;
-  color: white;
-  margin-left:30px;
-  font-weight:600;
-  color:#000;
+  border-radius: 3px;
+  width: 200px;
+  padding: 10px;
+  border: 1px solid #e2e0e0;
+  border-radius: 6px;
+  background-color: #cde8ac;
+  font-size: 20px;
+  color: black;
+  font-weight: 400;
+  cursor: pointer;
+  margin-top: 30px;
+  margin-left: 20px;
 `;
